+++ src/modules/image.js
@@ -6,6 +6,10 @@ const Module = Quill.import("core/module");
 class ImageBlot extends Image {
   static create(value) {
     let node = super.create(typeof value === "string" ? value : value.url);
+    if (value.style) {
+      node.setAttribute("style", value.style);
+    }
+
     if (value.latex) {
       node.dataset.latex = value.latex;
     }
@@ -17,6 +21,7 @@ class ImageBlot extends Image {
     return {
       url: node.getAttribute("src"),
       latex: node.dataset.latex,
+      style: node.style.cssText,
     };
   }
 }